@@ -57,56 +57,53 @@
     "tslib": "^2.8.1"
   },
   "devDependencies": {
-    "@babel/core": "~7.22.20",
-    "@babel/preset-env": "~7.22.20",
-    "@babel/preset-react": "~7.22.15",
-    "@babel/preset-typescript": "~7.22.15",
-    "@commitlint/cli": "^19.6.0",
-    "@commitlint/config-conventional": "^19.6.0",
-    "@commitlint/cz-commitlint": "^19.5.0",
-    "@rollup/plugin-commonjs": "^25.0.8",
-    "@rollup/plugin-node-resolve": "^15.3.0",
-    "@rollup/plugin-typescript": "^11.1.6",
+    "@babel/core": "^7.26.9",
+    "@babel/preset-env": "^7.26.9",
+    "@babel/preset-react": "^7.26.3",
+    "@babel/preset-typescript": "^7.26.0",
+    "@commitlint/cli": "^19.8.0",
+    "@commitlint/config-conventional": "^19.8.0",
+    "@commitlint/cz-commitlint": "^19.8.0",
+    "@rollup/plugin-commonjs": "^28.0.3",
+    "@rollup/plugin-node-resolve": "^16.0.0",
+    "@rollup/plugin-typescript": "^12.1.2",
     "@testing-library/jest-dom": "^6.6.3",
-    "@testing-library/react": "^14.3.1",
+    "@testing-library/react": "^16.2.0",
     "@types/jest": "^29.5.14",
-    "@types/react": "^18.3.12",
-    "@types/react-dom": "^18.3.1",
+    "@types/react": "^18.3.18",
+    "@types/react-dom": "^18.3.5",
     "@typescript-eslint/eslint-plugin": "^7.18.0",
     "@typescript-eslint/parser": "^7.18.0",
-    "antd": "^5.22.2",
-    "autoprefixer": "^10.4.20",
+    "antd": "^5.24.3",
+    "autoprefixer": "^10.4.21",
     "babel-jest": "^29.7.0",
     "commitizen": "^4.3.1",
     "create-puzzle": "^3.0.2",
-    "doly-icons": "^1.5.5",
-    "dumi": "^2.0.0",
+    "doly-icons": "^1.6.0",
+    "dumi": "^2.4.18",
     "eslint": "^8.57.1",
-    "eslint-plugin-react": "^7.37.2",
+    "eslint-plugin-react": "^7.37.4",
     "eslint-plugin-react-hooks": "^4.6.2",
     "husky": "^9.1.7",
-    "inquirer": "^9.3.7",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
-    "less": "^4.2.1",
-    "lint-staged": "^15.2.10",
-    "postcss": "^8.4.49",
+    "less": "^4.2.2",
+    "lint-staged": "^15.4.3",
+    "postcss": "^8.5.3",
     "postcss-css-variables": "^0.19.0",
-    "prettier": "^3.4.0",
+    "prettier": "^3.5.3",
     "prettier-plugin-two-style-order": "^1.0.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "rollup": "^4.27.4",
+    "rollup": "^4.35.0",
     "rollup-plugin-postcss": "^4.0.2",
-    "stylelint": "^14.16.1",
-    "stylelint-config-css-modules": "~4.2.0",
-    "stylelint-config-prettier": "^9.0.5",
-    "stylelint-config-rational-order": "^0.1.2",
-    "stylelint-config-standard": "^28.0.0",
+    "stylelint": "^16.15.0",
+    "stylelint-config-css-modules": "^4.4.0",
+    "stylelint-config-standard": "^37.0.0",
     "stylelint-declaration-block-no-ignored-properties": "^2.8.0",
-    "stylelint-no-unsupported-browser-features": "^5.0.4",
-    "typescript": "^5.7.2",
-    "ut2": "^1.14.0"
+    "stylelint-no-unsupported-browser-features": "^8.0.4",
+    "typescript": "^5.8.2",
+    "ut2": "^1.16.0"
   },
   "peerDependencies": {
     "react": ">=16.9.0"
