@@ -1,9 +1,7 @@
-module.exports = {
+export default {
   extends: [
     'stylelint-config-standard',
     'stylelint-config-css-modules',
-    'stylelint-config-rational-order',
-    'stylelint-config-prettier',
     'stylelint-no-unsupported-browser-features'
   ],
   plugins: ['stylelint-declaration-block-no-ignored-properties'],
